+++ DependencyInjection/Configuration.php
@@ -13,6 +13,7 @@ namespace Chaplean\Bundle\DtoHandlerBundle\DependencyInjection;
 
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
+use Symfony\Component\HttpFoundation\Response;
 
 /**
  * This class contains the configuration information for the bundle.
@@ -36,8 +37,21 @@ final class Configuration implements ConfigurationInterface
                     ])
                     ->prototype('scalar')->end()
                 ->end()
-            ->end()
-        ->end();
+                ->arrayNode('http_code_validation_groups')
+                    ->info('Validate DTO with the the group and throw a HTTP exception with the mentioned status code in case of violations')
+                    ->defaultValue([
+                        'validation_group' => 'Default',
+                        'http_status_code' => Response::HTTP_BAD_REQUEST,
+                    ])
+                    ->arrayPrototype()
+                        ->children()
+                            ->scalarNode('validation_group')->defaultNull()->end()
+                            ->integerNode('http_status_code')->defaultNull()->end()
+                            ->integerNode('priority')->defaultValue(0)->end()
+                        ->end()
+                    ->end()
+                ->end()
+            ->end();
 
         return $treeBuilder;
     }