+++ EventListener/DataTransferObjectValidationExceptionSubscriber.php
@@ -19,7 +19,7 @@ class DataTransferObjectValidationExceptionSubscriber implements EventSubscriber
     /**
      * @return array
      */
-    public static function getSubscribedEvents()
+    public static function getSubscribedEvents(): array
     {
         return [
             KernelEvents::EXCEPTION => 'onKernelException'